ai_backup/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
httpx==0.27.0
google-generativeai==0.6.0
python-dotenv==1.0.1
